Update for 115
Since TB102, the Service module is globally available for extensions and does not need to be loaded anymore.

The promptUsernameAndPassword() function still exists in TB115, but throws an error that it does not exists. Switching to asyncPromptUsernameAndPassword (which is TB115 only).

I added a fallback for TB102, because there is a strange bug on ATN, that add-ons that have their latest submission being 115 only, not showing up in 102 anymore (or marked as incompatible). So I tried to create a version which runs in both.